This commit lays out the Equation and Term classes for the shallow water equations.
The motivation is to allow for greater flexibility with timestepping
schemes, the intention being that terms can be labelled according to
whether they should be evaluated explicitly or implicitly.

At the moment, this only reproduces the current split forcing and
advection timestepping scheme. The advection term is detected as being
an instance of the TransportTerm class - this would be replaced by
term labels once they are implemented.

The Equation class takes in the function space and defines test and
trial functions and a mass term. It contains a dictionary of terms
where the key is the name of the class defining the term and the value
is the term class. It defines the mass term, which can be evaluated on
a trial function or a function. Terms are added by calling the
add_term method, passing in the term class. The Equation class also
has a __call__ method which requires arguments q and fields where q is
the thing that the terms are evaluated on and the fields is an
instance of the FieldCreator class that contains the values of the
fields at the correct time levels required to evalute the term.

So far the only Equation subclasses implemented are:

- AdvectionEquation (only works for DG advection, not embedded DG or SUPG)
- ShallowWaterMomentumEquation (doesn't yet work for the EulerPoincare option)
- ShallowWaterDepthEquation

These equations are set up via the Equations class. The only Equations
subclass currently implemented is ShallowWaterEquations. Equations
classes should define the following properties: the fieldlist, the
equation_list (a list of instances of equation classes corresponding
to the fields in the fieldlist) and the mixed_function_space. They
should also define the method for creating the function
spaces. Calling ShallowWaterEquations(field) will return the equation
for that field.

The Term class takes in state and saves the physical parameters. It
constructs some things commonly used in evaluating terms (at the
moment just the FacetNormal but I envisage putting things like the
vertical normal and the perp operator in here, rather than in
state). Each Term subclass must define an evaluate method that takes
in the test function, q and fields and evaluates the term - this is
called via the Equation __call__ method so Equation passes in the test
function. The Term class has a default off_centering parameter of 0.5
(used in the construction of the explicit and implicit forcing
solvers) - this should be exposed to the user so that it can be
altered.

Other necessary changes have been made to the state class, the forcing
and advection classes and the timestepping classes:

- the fieldlist and function space generation has been moved to the
  Equations class

- the FieldCreator class is now used everywhere that fields are
  allocated and has been rewritten to avoid the need to create
  dictionaries so that individual fields can be accessed by name

- the _allocate_state method of the State class is not used; instead
  the fields at different time levels are allocated from the
  timestepping scheme. This is because different timestepping schemes
  require a different number of stages. For example, the current split
  scheme needs xstar as the output of the initial forcing stage but
  for an implicit midpoint scheme that does not separate the forcing
  and advection, this is not needed.

- the forcing class takes in the equations and constructs the explicit
  and implicit solvers, omitting any terms that are instances of the
  TransportTerm - in future this will be replaced by term labels. This
  makes the forcing class generic - there is no need for specific
  subclasses for each set of equations.

- in the BaseTimestepping class the semi_implicit_step abstract method
  has been replaces with an abstract method called timestep so that it
  can also be used for explicit timestepping. These is a new
  BaseSemiImplicitTimestepper class that requires subclasses to define
  a semi_implicit_step method.